int instead of long for mxArray dimensions to avoid warnings

+++ vpp_init_input.c
@@ -20,7 +20,7 @@ void mexFunction(int nlhs, mxArray *plhs[],
     mxFree((void*) filename);
 
     if (file) {
-        long dims[] = {4};
+        int dims[] = {4};
         plhs[0] = mxCreateNumericArray(1, dims, mxUINT64_CLASS, mxREAL);
         UINT64_T* pr = (UINT64_T *) mxGetData(plhs[0]);
         pr[1] = w;
